Safely allow requires inside of conditional directives
If a module is imported within a conditional preprocessor directive,
e.g. \#ifdef, it will be treated as a separate module to avoid it
being deduped out of existence if modules are required in the wrong
order.

A caveat is that this causes shader files to be larger. It would be
nice to be able to rearrange dependencies to avoid this issue instead,
but I'm not sure on the best approach for it. The approach included
here is safe and does not require restructuring the code too much.

Closes #7